requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic[email]
httpx
python-jose[cryptography]
passlib[bcrypt]
websockets
pandas
numpy
python-dateutil
pytest
pytest-asyncio
pytest-cov
pytest-mock
tenacity
cryptography>=42.0.0
SQLAlchemy
pandas-ta 